add Zone status (#2)
Query the status field of the zone object use to report the
Spectrum config. Cloudflare allows defining zones without
checking for uniqueness of the referenced domain names; this
repo could therefore be set to report on a moved or deactivated
zone. The report should have zone_status: active to show the
zone describes the configuration currently serving traffic.

Signed-off-by: Ralph Giles <rgiles@brave.com>